backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
bcrypt
passlib[bcrypt]
python-jose[cryptography]